# py-scikit-learn: add v1.5.0