--- a/var/spack/repos/builtin/packages/py-scikit-learn-extra/package.py
+++ b/var/spack/repos/builtin/packages/py-scikit-learn-extra/package.py
@@ -27,4 +27,5 @@ class PyScikitLearnExtra(PythonPackage):
     depends_on("py-cython@0.28.5:", type="build")
     depends_on("py-numpy@1.13.3:", type=("build", "run"))
     depends_on("py-scipy@0.19.1:", type=("build", "run"))
-    depends_on("py-scikit-learn@0.23:", type=("build", "run"))
+    # For upperbound see https://github.com/scikit-learn-contrib/scikit-learn-extra/issues/171
+    depends_on("py-scikit-learn@0.23:1.2", type=("build", "run"))