Sklearn has also ported to meson! Not as much of a build speedup though:

### Before

```
  Stage: 2.26s.  Install: 56.53s.  Post-install: 0.34s.  
 59.25s
```

### After

```
  Stage: 2.29s.  Install: 47.79s.  Post-install: 0.35s.  Total: 50.55s
```

https://scikit-learn.org/stable/whats_new/v1.5.html