@rgommers 2024-05-21
> Not as much of a build speedup though:

I guess because their build parallelized better with `setuptools`. Still, I had expected a bit more gain.