@rgommers 2024-05-21
The recipe for scikit-learn is refreshingly simple:) 

In less good news, the first job I checked (https://gitlab.spack.io/spack/spack/-/jobs/11339913) claims it passes but does contain a failure:
```  
FAILED: PyScikitLearnExtra::test_imports_sklearn_extra.robust: Command exited with status 1:
```