@tldahlgren 2024-05-21
> @tldahlgren do you know anything about this? Are the import tests allowed to fail, or are they not raising an error like they should?

Stand-alone (including import) tests are currently allowed to fail during CI since they are being executed only on successful build since a key goal of the build job is to populate the build cache.  This will change once the tests are split into a separate job (e.g., resurrect https://github.com/spack/spack/pull/33180).