@adamjstewart 2024-05-23
Fixed the import error @rgommers reported. Looks like py-scikit-learn-extra doesn't support newer py-scikit-learn. In fact, it hasn't been updated in quite a while. We may want to consider removing it from CI.